+++ inc/Score.hpp
@@ -8,11 +8,11 @@ public:
   Score() : m_score(0), m_level(1) {}
 
   void update_score(const unsigned no_of_lines);
-  unsigned get_score() const { return m_score; }
+  unsigned long get_score() const { return m_score; }
   unsigned get_level() const { return m_level; }
 
 private:
-  unsigned m_score;
+  unsigned long m_score;
   unsigned m_level;
 };
